requirements.txt
torch==2.9.0
torchvision==0.24.0
matplotlib==3.10.0
Pillow==11.3.0
mlflow==3.8.1
pytest==9.0.2